% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{package}
\name{hordr}
\alias{hordr}
\alias{hordr-package}
\title{Higher order functions and function wrappers to empower and simplify common operations.}
\description{
Hordr adds features, generalizes, and fills in gaps in standards R operations to add power and flexibility.
}

