% Generated by roxygen2 (4.0.0): do not edit by hand
\name{pb}
\alias{pb}
\title{A wrapper that will at a progress bar to many higher order functions}
\usage{
pb(fun)
}
\arguments{
\item{fun}{the function to evaluate}

\item{...}{all other arguments to be passed to fun}
}
\description{
This process time utility can be wrapped around any normal function with no effect other than printing the process time to the screen.  This allows visual feedback on timing without otherwise interrupting the operation of the function.
}
\examples{
results<-pb(lapply)(sqrt,rep.int(x=5,times=1000000))
results<-pb(Reduce)('+',sqrt(c(1:100000)))
}

