% Generated by roxygen2 (4.0.0): do not edit by hand
\name{rollApply}
\alias{rollApply}
\title{Applies a function to a rolling window along any data object.}
\usage{
rollApply(data, fun, window = len(data), minimum = 1, align = "left")
}
\arguments{
\item{fun}{the function to evaluate}

\item{...}{all other arguments to be passed to fun}
}
\description{
Similar but not identical to \code{\link{rollapply}} in package {\code\link{zoo}}, with the advantage that it works on any type of data structure (vector, list, matrix, etc) instead of requiring a \code{\link{zoo}} object.
}
\examples{
rollApply(1:100,sum,minimum=2,window=2)
rollApply(1:100,sum,minimum=2,window=2,align='right')
}

