% Generated by roxygen2 (4.0.0): do not edit by hand
\name{time}
\alias{time}
\title{An alternative to \code{\link{system.time}} that prints the times to the screen while returning the result of the original operation.}
\usage{
time(fun)
}
\arguments{
\item{fun}{the function to evaluate}

\item{...}{all other arguments to be passed to fun}
}
\description{
This process time utility can be wrapped around any normal function with no effect other than printing the process time to the screen.  This allows visual feedback on timing without otherwise interrupting the operation of the function.
}
\examples{
results<-time (vectorize(sum))(rep.int(x=5,times=1000000))
results
}

